chore: Move catalog into iceberg
