+++ crates/data_components/src/iceberg/catalog/mod.rs
@@ -14,4 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-pub mod catalog;
+pub mod hadoop;
+pub mod rest;
