+++ crates/data_components/src/iceberg/catalog/rest/mod.rs
@@ -14,4 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-pub mod catalog;
+mod catalog;
+pub use catalog::*;
+
+mod provider;
+pub use provider::*;
